Refactor browser_manager profile_manager proxy_manager modules
